Changed the slider to have negative clicks

+++ GAME01_IGME450/Assets/Scripts/Painting Scripts/Painting.cs
@@ -24,7 +24,7 @@ public class Painting : MonoBehaviour
     public void Upvote(float multiplier, GameObject golfbar)
     {
         clicks += 1;
-        int golfMulti = golfbar.GetComponent<GolfMeter>().GetBonus(clicks);
+        float golfMulti = golfbar.GetComponent<GolfMeter>().GetBonus(clicks);
         HandleClick(multiplier, 1, golfMulti);
 
         if (popularity > 100)
@@ -34,14 +34,14 @@ public class Painting : MonoBehaviour
     public void Downvote(float multiplier, GameObject golfbar)
     {
         clicks += 1;
-        int golfMulti = golfbar.GetComponent<GolfMeter>().GetBonus(clicks);
+        float golfMulti = golfbar.GetComponent<GolfMeter>().GetBonus(clicks);
         HandleClick(multiplier, -1, golfMulti);
 
         if (popularity < 0)
             popularity = 0;
     }
 
-    private void HandleClick(float multiplier, int sign, int _golfMulti)
+    private void HandleClick(float multiplier, int sign, float _golfMulti)
     {
 
         if (multiplier > 0)
